Coldown automatique et en floatant

+++ src/main/java/shionn/hexas/bot/messages/Message.java
@@ -231,8 +231,8 @@ public class Message {
 		return substitution("po", po);
 	}
 
-	public Message coldDown(int coldDown) {
-		substitution.put("coldDown", Integer.toString(coldDown));
+	public Message coldDown(float coldDown) {
+		substitution.put("coldDown", Float.toString(coldDown));
 		return this;
 	}
